+++ src/redux/actions/confirm-unassign-agreement.js
@@ -1,0 +1,5 @@
+export const CONFIRM_UNASSIGN_AGREEMENT = 'CONFIRM_UNASSIGN_AGREEMENT';
+
+export const confirmUnassignAgreement = () => ({
+  type: CONFIRM_UNASSIGN_AGREEMENT,
+});