+++ src/redux/actions/get-agreement-lines-failure.js
@@ -1,0 +1,6 @@
+export const GET_AGREEMENT_LINES_FAILURE = 'GET_AGREEMENT_LINES_FAILURE';
+
+export const getAgreementLinesFailure = payload => ({
+  type: GET_AGREEMENT_LINES_FAILURE,
+  payload,
+});