+++ src/redux/actions/unassign-agreement.js
@@ -1,0 +1,6 @@
+export const UNASSIGN_AGREEMENT = 'UNASSIGN_AGREEMENT';
+
+export const unassignAgreement = payload => ({
+  type: UNASSIGN_AGREEMENT,
+  payload,
+});