+++ test/bigtest/interactors/agreements-accordion.js
@@ -9,7 +9,10 @@ import {
 
 export default @interactor class {
   isExpanded = !!attribute('#accordion-toggle-button-packageShowAgreements', 'aria-expanded');
-  agreements = collection('[data-test-agreements-list-item]');
+  agreements = collection('[data-test-agreements-list-item]', {
+    clickTrashIcon: clickable('[data-test-delete-agreement]'),
+  });
+
   hasNewButton = isPresent('[data-test-new-button]');
   clickNewButton = clickable('[data-test-new-button]');
   clickSection = clickable('[class^=defaultCollapseButton]');