+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 * Fixed View Package > Find Title - Accordions missing some ARIA attributes. (UIEH-925)
 * Fix bug with `<Datepicker>` in Resource Edit Page. (UIEH-961)
 * Fix issue in agreements accordion with displaying start date. (UIEH-971)
+* Ability to unlink an agreement from package/resource. (UIEH-956)
 
 ## [4.0.1] (https://github.com/folio-org/ui-eholdings/tree/v4.0.0) (2020-07-08)